Move roomname generator to React
