+++ app.js
@@ -24,7 +24,9 @@ const LogCollector = Logger.LogCollector;
 import JitsiMeetLogStorage from "./modules/util/JitsiMeetLogStorage";
 
 import URLProcessor from "./modules/config/URLProcessor";
-import { generateRoomWithoutSeparator } from './modules/util/RoomnameGenerator';
+import {
+    generateRoomWithoutSeparator
+} from './react/features/base/util/roomnameGenerator';
 
 import UI from "./modules/UI/UI";
 import settings from "./modules/settings/Settings";
