+++ react/features/base/util/index.js
@@ -1,1 +1,2 @@
 export * from './loadScript';
+export * from './roomnameGenerator';
